requirements.txt
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.42.0
djangorestframework==3.12.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
Pyphen==0.9.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
